# feat: NSA training breakthrough - stable A100 production at 39 toks/s with optimization plan